Make enabled semantics more well-defined with respect to reset operations
When resetting the active config (with or without supplying a mask), and the
config message type has a field named 'enabled', the server will create a new
revision and preserve history, instead of deleting the object from the
active store. When there is no 'enabled' field in the message type, it
will delete the object as usual.

Add revision field to generic history request

Add unit tests for the base config server