Sort files before processing in cli generator
This prevents flag sets from being ordered non-deterministically in
the generated code in certain cases. Specifically, if two separate
messages have flag set generators triggered by dependent messages in
separate files, the order of those dependent files in the input list
can cause the order of the flag sets in the generated code to change.
The input list is not guaranteed to be sorted exactly the same way
each time the code generator is run; it only guarantees that the
elements are in _a_ topological order.