fix: merge of the configuration with empty list.
For some reason the configuration merge function is called with the src
and dest "inverted". In other words, the dest map contains the defaults
values defined as the first file in the -c command line argument of the
command. As result, the merge function fails because it detects this as
incompatible types to be merged. This commit fixes that by removing the
slice len check.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>